rename package to catalog-sort-price
