+++ bundles/catalog-sort-price/README.md
@@ -1,0 +1,10 @@
+# catalog sort price
+[![license](https://img.shields.io/github/license/fond-of-kudu/catalog-sort-price.svg)](https://packagist.org/packages/fond-of-kudu/catalog-sort-price)
+
+## What it does
+
+## Installation
+
+```
+composer require fond-of-kudu/catalog-sort-price
+```
