+++ bundles/catalog-sort-price/composer.json
@@ -1,5 +1,5 @@
 {
-  "name": "fond-of-kudu/catalog-sorting-price",
+  "name": "fond-of-kudu/catalog-sort-price",
   "description": "n/a",
   "license": "MIT",
   "authors": [
