+++ bundles/catalog-sort-price/src/FondOfKudu/Client/CatalogSortPrice/CatalogSortPriceClient.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace FondOfKudu\Client\CatalogSortPrice;
+
+use Spryker\Client\Kernel\AbstractClient;
+
+/**
+ * @method \FondOfKudu\Client\CatalogSortPrice\CatalogSortPriceFactory getFactory()
+ */
+class CatalogSortPriceClient extends AbstractClient implements CatalogSortPriceClientInterface
+{
+}
