+++ bundles/catalog-sort-price/src/FondOfKudu/Client/CatalogSortPrice/CatalogSortPriceClientInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace FondOfKudu\Client\CatalogSortPrice;
+
+interface CatalogSortPriceClientInterface
+{
+}
