+++ bundles/catalog-sort-price/src/FondOfKudu/Client/CatalogSortPrice/CatalogSortPriceFactory.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace FondOfKudu\Client\CatalogSortPrice;
+
+use Spryker\Client\Kernel\AbstractFactory;
+
+class CatalogSortPriceFactory extends AbstractFactory
+{
+}
