+++ bundles/catalog-sort-price/src/FondOfKudu/Shared/CatalogSortPrice/CatalogSortPriceConstants.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace FondOfKudu\Shared\CatalogSortPrice;
+
+interface CatalogSortPriceConstants
+{
+}
