--- bundles/catalog-sorting-price/README.md
@@ -1,10 +1,0 @@
-# catalog sorting price
-[![license](https://img.shields.io/github/license/fond-of-kudu/catalog-sorting-price.svg)](https://packagist.org/packages/fond-of-kudu/catalog-sorting-price)
-
-## What it does
-
-## Installation
-
-```
-composer require fond-of-kudu/catalog-sorting-price
-```
