--- bundles/catalog-sorting-price/src/FondOfKudu/Client/CatalogSortingPrice/CatalogSortingPriceClient.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace FondOfKudu\Client\CatalogSortingPrice;
-
-use Spryker\Client\Kernel\AbstractClient;
-
-/**
- * @method \FondOfKudu\Client\CatalogSortingPrice\CatalogSortingPriceFactory getFactory()
- */
-class CatalogSortingPriceClient extends AbstractClient implements CatalogSortingPriceClientInterface
-{
-}
