--- bundles/catalog-sorting-price/src/FondOfKudu/Client/CatalogSortingPrice/CatalogSortingPriceClientInterface.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace FondOfKudu\Client\CatalogSortingPrice;
-
-interface CatalogSortingPriceClientInterface
-{
-}
