--- bundles/catalog-sorting-price/src/FondOfKudu/Client/CatalogSortingPrice/CatalogSortingPriceFactory.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace FondOfKudu\Client\CatalogSortingPrice;
-
-use Spryker\Client\Kernel\AbstractFactory;
-
-class CatalogSortingPriceFactory extends AbstractFactory
-{
-}
