--- bundles/catalog-sorting-price/src/FondOfKudu/Shared/CatalogSortingPrice/CatalogSortingPriceConstants.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace FondOfKudu\Shared\CatalogSortingPrice;
-
-interface CatalogSortingPriceConstants
-{
-}
