+++ dandelion.json
@@ -7,7 +7,7 @@
       "version": "1.2.2"
     },
     "catalog-sorting-price": {
-      "path": "bundles/catalog-sorting-price",
+      "path": "bundles/catalog-sort-price",
       "version": "1.0.0-beta.1"
     },
     "checkout-rest-api": {
